-switched fields from TextField to TextArea class -added height increase for fields upon focus -adjusted PresetSetupTab ui -added remove buttons for PresetSetupTab tables -upon startup a "saves" folder is created for json settings storage -added tooltips to most ui elements -adjusted entry saving and loading logic

+++ src/main/java/org/ashot/microservice_starter/data/icon/Icons.java
@@ -108,7 +108,7 @@ public class Icons {
         return icon;
     }
 
-    public static Glyph getBrowseIcon(int size) {
+    public static Glyph getBrowseIcon(double size) {
         Glyph icon = fontAwesome.create(FontAwesome.Glyph.SEARCH);
         icon.size(size);
         icon.color(Color.DARKSLATEBLUE);
